api/requirements.txt
fastapi[all]>=0.78.0
uvicorn[standard]==0.17.6
pytest==7.2.1
psycopg[binary,pool]>=3.1.2
jwtdown-fastapi>=0.2.0
